feat(qb): SJRA-1371 rename query builder instead of v3
