+++ frontend/components/base/query-builder/facets/search-facet.tsx
