+++ frontend/components/base/query-builder/libs/aggregations.ts
