+++ frontend/components/base/query-builder/pills/search-query-pill.tsx
