+++ frontend/components/base/query-builder/query-builder-skeleton.tsx
