+++ frontend/components/base/query-builder/query-builder.tsx
@@ -3,15 +3,15 @@ import { X } from 'lucide-react';
 import useSWR from 'swr';
 
 import { SavedFilter, SavedFilterType } from '@/api/index';
-import { SidebarGroups } from '@/components/base/query-builder-v3//facets/sidebar-groups';
-import { FacetList } from '@/components/base/query-builder-v3/facets/facet-list';
-import { FacetConfigContext } from '@/components/base/query-builder-v3/facets/hooks/use-facet-config';
-import { getAggregationsFetcher } from '@/components/base/query-builder-v3/hooks/use-aggregation-builder';
-import { IQBContext, IQBFetcher, QBProvider } from '@/components/base/query-builder-v3/hooks/use-query-builder';
-import { useQueryBuilderGetPreferenceEffect } from '@/components/base/query-builder-v3/hooks/use-query-builder-preference';
-import { getVisibleAggregations } from '@/components/base/query-builder-v3/libs/aggregations';
-import QueriesBarCard from '@/components/base/query-builder-v3/queries-bar-card';
-import { QueryBuilderSkeleton } from '@/components/base/query-builder-v3/query-builder-skeleton';
+import { SidebarGroups } from '@/components/base/query-builder//facets/sidebar-groups';
+import { FacetList } from '@/components/base/query-builder/facets/facet-list';
+import { FacetConfigContext } from '@/components/base/query-builder/facets/hooks/use-facet-config';
+import { getAggregationsFetcher } from '@/components/base/query-builder/hooks/use-aggregation-builder';
+import { IQBContext, IQBFetcher, QBProvider } from '@/components/base/query-builder/hooks/use-query-builder';
+import { useQueryBuilderGetPreferenceEffect } from '@/components/base/query-builder/hooks/use-query-builder-preference';
+import { getVisibleAggregations } from '@/components/base/query-builder/libs/aggregations';
+import QueriesBarCard from '@/components/base/query-builder/queries-bar-card';
+import { QueryBuilderSkeleton } from '@/components/base/query-builder/query-builder-skeleton';
 import { SidebarProvider } from '@/components/base/shadcn/sidebar';
 import { AggregationConfig, ApplicationId, AppsConfig, useConfig } from '@/components/cores/applications-config';
 import { cn } from '@/components/lib/utils';
