+++ frontend/components/base/query-builder/saved-filter/delete-filter-button.tsx
