+++ frontend/components/base/query-builder/saved-filter/discard-filter-button.tsx
